requirements.txt
python-can>=4.5.0
cantools>=39.4.0
PyYAML>=6.0.1
typer>=0.12.3
rich>=13.8.1

# testing
pytest>=8.3.0
